requirements.txt
paho-mqtt>=1.6.1
azure-cognitiveservices-speech>=1.31.0
pyttsx3>=2.90
openai>=1.3.0
python-dotenv>=1.0.0
PyAudio>=0.2.13 